Fix hp friend fetch

+++ lib/view_model/friend_data_notifier.dart
@@ -42,9 +42,9 @@ class FriendData extends _$FriendData {
     friendDataDict.forEach((key, value) {
       add(FriendState(
           friendName: friendIdDict[key],
-          currentHP: value["friend_hp"],
+          currentHP: value["friend_hp"].toInt(),
           avatarUrl: value["friend_avatar_image"],
-          hpFontColor: selectHPfontColor(value["friend_hp"])));
+          hpFontColor: selectHPfontColor(value["friend_hp"].toInt())));
     });
   }
